isGlobalHandler
