+++ packages/stentor-constants/src/requests.ts
@@ -1,0 +1,19 @@
+/*! Copyright (c) 2020, XAPPmedia */
+import {
+    AttachmentRequestID,
+    InputUnknownID,
+    KnowledgeAnswerID,
+    LaunchRequestID,
+    OptionSelectID,
+    PermissionGrantID,
+    SignInID
+} from "stentor-models";
+
+export const LAUNCH_REQUEST_ID: LaunchRequestID = "LaunchRequest";
+export const INPUT_UNKNOWN_ID: InputUnknownID = "InputUnknown";
+
+export const PERMISSION_GRANT_ID: PermissionGrantID = "PermissionGrant";
+export const OPTION_SELECT_ID: OptionSelectID = "OptionSelect";
+export const SIGN_IN_ID: SignInID = "SignIn";
+export const KNOWLEDGE_ANSWER_ID: KnowledgeAnswerID = "KnowledgeAnswer";
+export const ATTACHMENT_REQUEST_ID: AttachmentRequestID = "AttachmentRequest";
