+++ packages/stentor-guards/package.json
@@ -28,6 +28,7 @@
     "typescript": "3.8.3"
   },
   "dependencies": {
+    "stentor-constants": "1.23.34",
     "stentor-utils": "1.23.34"
   },
   "peerDependencies": {
