+++ packages/stentor-handler/src/AbstractHandler/Handler.ts
@@ -1,4 +1,5 @@
 /*! Copyright (c) 2019, XAPPmedia */
+import { INPUT_UNKNOWN_ID } from "stentor-constants";
 import { isActionable } from "stentor-guards";
 import { REPEAT_INTENT } from "stentor-interaction-model";
 import { log } from "stentor-logger";
@@ -16,7 +17,7 @@ import {
     SlotTypeMap,
     UserDataType
 } from "stentor-models";
-import { INPUT_UNKNOWN_ID, keyFromRequest } from "stentor-request";
+import { keyFromRequest } from "stentor-request";
 import { determineResponse, getResponse } from "stentor-response";
 import { manipulateStorage } from "stentor-storage";
 import { concatText, dessmlify, findValueForKey } from "stentor-utils";
